feat : バックエンドに合わせて取得したデータを扱えるように変更、またUIを一部改善 feature #233
