+++ frontend/src/app/routes/app/management/account/accountEdit.tsx
@@ -1,6 +1,10 @@
 import { useEffect } from "react";
 import { useNavigate, useParams } from "react-router";
 import { useAccount } from "@/features/management/hooks/useAccount";
+import {
+  useModifyStudent,
+  useModifyTeacher,
+} from "@/features/management/hooks/useAccountMutations";
 import { BackPage } from "@/features/management/components/backPage/backPage";
 import {
   StudentAccountEditForm,
@@ -12,6 +16,8 @@ import {
 } from "@/features/management/types/account";
 import styles from "@/features/management/style/accountEdit.module.css";
 import { paths } from "@/config/paths";
+import { useAuth } from "@/features/auth/hooks/useAuth";
+import { decodeUserParams } from "@/features/auth/utils/jwt";
 
 const isTeacherAccount = (
   account: StudentAccountEditType | TeacherAccountEditType,
@@ -24,6 +30,11 @@ const AccountEdit = () => {
   const { userId } = useParams();
   const parsedUserId = Number(userId) || 0;
 
+  const { authToken } = useAuth();
+  const userInfo = decodeUserParams(authToken);
+  const role = userInfo?.role ?? "STUDENT";
+  const isAdmin = role === "ADMIN_SCHOOL";
+
   useEffect(() => {
     if (!userId || !Number.isFinite(parsedUserId) || parsedUserId <= 0) {
       navigate(paths.app.management.account.list.path, { replace: true });
@@ -31,6 +42,48 @@ const AccountEdit = () => {
   }, [userId, navigate, parsedUserId]);
 
   const { data: account, isFetching, isError } = useAccount(parsedUserId);
+
+  const modifyStudent = useModifyStudent({
+    onSuccess: () => {
+      alert("更新しました");
+      navigate(paths.app.management.account.list.path, { replace: true });
+    },
+    onError: () => {
+      alert("更新に失敗しました");
+    },
+  });
+
+  const modifyTeacher = useModifyTeacher({
+    onSuccess: () => {
+      alert("更新しました");
+      navigate(paths.app.management.account.list.path, { replace: true });
+    },
+    onError: () => {
+      alert("更新に失敗しました");
+    },
+  });
+
+  const handleStudentSubmit = (formData: StudentAccountEditType) => {
+    modifyStudent.mutate({
+      userId: formData.userId,
+      name: formData.name,
+      mailAddress: formData.email,
+      graduateDate: formData.graduateDate,
+      accountStopFlag: !!formData.accountStopFlag,
+    });
+  };
+
+  const handleTeacherSubmit = (formData: TeacherAccountEditType) => {
+    modifyTeacher.mutate({
+      userId: formData.userId,
+      name: formData.name,
+      mailAddress: formData.email,
+      authorityFlag: !!formData.authority,
+      accountStopFlag: !!formData.accountStopFlag,
+    });
+  };
+
+  const isPending = modifyStudent.isPending || modifyTeacher.isPending;
 
   const handleDelete = async () => {
     if (!userId) {
@@ -57,9 +110,17 @@ const AccountEdit = () => {
       ) : isFetching || !account ? (
         <div>読み込み中...</div>
       ) : isTeacherAccount(account) ? (
-        <TeacherAccountEditForm {...account} />
+        <TeacherAccountEditForm
+          {...account}
+          onSubmit={handleTeacherSubmit}
+          isPending={isPending}
+        />
       ) : (
-        <StudentAccountEditForm {...account} />
+        <StudentAccountEditForm
+          {...account}
+          onSubmit={handleStudentSubmit}
+          isPending={isPending}
+        />
       )}
 
       <div className={styles.actions}>
@@ -67,16 +128,19 @@ const AccountEdit = () => {
           form="accountEditForm"
           className={`${styles.button} ${styles.primary}`}
           type="submit"
-        >
-          更新
+          disabled={isPending}>
+          {isPending ? "更新中..." : "更新"}
         </button>
-        <button
-          className={`${styles.button} ${styles.danger}`}
-          type="button"
-          onClick={handleDelete}
-        >
-          削除
-        </button>
+
+        {isAdmin && (
+          <button
+            className={`${styles.button} ${styles.danger}`}
+            type="button"
+            onClick={handleDelete}
+            disabled={isPending}>
+            削除
+          </button>
+        )}
       </div>
     </div>
   );
