+++ frontend/src/features/management/components/accountEditForm/accountEditForm.module.css
@@ -7,6 +7,15 @@
   flex-shrink: 0;
 }
 
+.editForm fieldset {
+  all: unset;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  width: 100%;
+  min-width: 0;
+}
+
 .row {
   display: flex;
   flex-direction: row;
@@ -88,7 +97,9 @@
   position: relative;
   cursor: pointer;
   flex: 0 0 auto;
-  transition: background-color 120ms ease, border-color 120ms ease;
+  transition:
+    background-color 120ms ease,
+    border-color 120ms ease;
 }
 
 .toggle input[type="checkbox"]::before {
