+++ frontend/src/features/management/components/accountRegisterTable/accountRegisterTable.tsx
@@ -70,12 +70,11 @@ export const StudentAccountRegisterTable = ({
         <table className={styles.accountTable}>
           <thead>
             <tr>
+              <th>ID</th>
               <th>名前</th>
               <th>学年</th>
               <th>メールアドレス</th>
               <th>パスワード</th>
-              <th>入学時期</th>
-              <th>卒業見込み</th>
               <th className={styles.actionsHead}>操作</th>
             </tr>
           </thead>
@@ -86,8 +85,8 @@ export const StudentAccountRegisterTable = ({
                   key={index}
                   ref={(e) => {
                     tableRef.current[index] = e;
-                  }}
-                >
+                  }}>
+                  <td>{account.showUserId}</td>
                   <td>{account.name}</td>
                   <td>{account.grade}</td>
                   <td>{account.email}</td>
@@ -99,8 +98,7 @@ export const StudentAccountRegisterTable = ({
                           <button
                             type="button"
                             className={styles.actionButton}
-                            onClick={() => toggleConfig(index)}
-                          >
+                            onClick={() => toggleConfig(index)}>
                             <FaRegEye />
                           </button>
                         </>
@@ -110,32 +108,27 @@ export const StudentAccountRegisterTable = ({
                           <button
                             type="button"
                             className={styles.actionButton}
-                            onClick={() => toggleConfig(index)}
-                          >
+                            onClick={() => toggleConfig(index)}>
                             <FaRegEyeSlash />
                           </button>
                         </>
                       )}
                     </div>
                   </td>
-                  <td>{account.entryDate}</td>
-                  <td>{account.graduateDate}</td>
                   <td className={styles.actionsCell}>
                     <div className={styles.actionsInner}>
                       <button
                         type="button"
                         className={styles.actionButton}
                         title="編集"
-                        onClick={() => handleEdit(account, index)}
-                      >
+                        onClick={() => handleEdit(account, index)}>
                         <MdEdit color="rgb(56, 94, 164)" />
                       </button>
                       <button
                         type="button"
                         className={styles.actionButton}
                         title="削除"
-                        onClick={() => handleDelete(account)}
-                      >
+                        onClick={() => handleDelete(account)}>
                         <FaRegTrashAlt color="#fa5959" />
                       </button>
                     </div>
@@ -209,6 +202,7 @@ export const TeacherAccountRegisterTable = ({
         <table className={styles.accountTable}>
           <thead>
             <tr>
+              <th>ID</th>
               <th>名前</th>
               <th>メールアドレス</th>
               <th>パスワード</th>
@@ -222,8 +216,8 @@ export const TeacherAccountRegisterTable = ({
                   key={index}
                   ref={(e) => {
                     tableRef.current[index] = e;
-                  }}
-                >
+                  }}>
+                  <td>{account.showUserId}</td>
                   <td>{account.name}</td>
                   <td>{account.email}</td>
                   <td>
@@ -234,8 +228,7 @@ export const TeacherAccountRegisterTable = ({
                           <button
                             type="button"
                             className={styles.actionButton}
-                            onClick={() => toggleConfig(index)}
-                          >
+                            onClick={() => toggleConfig(index)}>
                             <FaRegEye />
                           </button>
                         </>
@@ -245,8 +238,7 @@ export const TeacherAccountRegisterTable = ({
                           <button
                             type="button"
                             className={styles.actionButton}
-                            onClick={() => toggleConfig(index)}
-                          >
+                            onClick={() => toggleConfig(index)}>
                             <FaRegEyeSlash />
                           </button>
                         </>
@@ -259,16 +251,14 @@ export const TeacherAccountRegisterTable = ({
                         type="button"
                         className={styles.actionButton}
                         title="編集"
-                        onClick={() => handleEdit(account, index)}
-                      >
+                        onClick={() => handleEdit(account, index)}>
                         <MdEdit color="rgb(56, 94, 164)" />
                       </button>
                       <button
                         type="button"
                         className={styles.actionButton}
                         title="削除"
-                        onClick={() => handleDelete(account)}
-                      >
+                        onClick={() => handleDelete(account)}>
                         <FaRegTrashAlt color="#fa5959" />
                       </button>
                     </div>
