+++ frontend/src/features/management/components/csvUploadField/csvUploadField.module.css
@@ -86,6 +86,8 @@
   width: 100%;
 }
 
+/* ── テンプレートダウンロード ── */
+
 .templateField {
   display: flex;
   flex-direction: column;
@@ -93,13 +95,54 @@
   justify-content: center;
   height: 100%;
   width: 25%;
-  gap: 3px;
+  gap: 6px;
 }
 
-.templateField label {
-  line-height: 1.3;
-  color: #666666;
+.templateLabel {
+  font-size: 0.75rem;
+  font-weight: 600;
+  color: #64748b;
+  letter-spacing: 0.04em;
+  text-transform: uppercase;
 }
+
+.templateLink {
+  display: inline-flex;
+  align-items: center;
+  gap: 6px;
+  padding: 6px 14px;
+  border-radius: 6px;
+  background-color: rgb(50, 85, 150);
+  color: #fff;
+  font-size: 0.85rem;
+  font-weight: 600;
+  text-decoration: none;
+  white-space: nowrap;
+  transition: background-color 0.15s ease;
+}
+
+.templateLink:hover {
+  background-color: rgb(40, 70, 130);
+}
+
+.templateLink:active {
+  background-color: rgb(35, 60, 115);
+}
+
+.templateIcon {
+  font-size: 0.8rem;
+  flex-shrink: 0;
+}
+
+.templateNote {
+  font-size: 0.7rem;
+  color: #94a3b8;
+  text-align: center;
+  line-height: 1.4;
+  margin: 0;
+}
+
+/* ── レスポンシブ ── */
 
 @media (max-width: 960px) {
   .uploadContainer {
@@ -150,14 +193,14 @@
 
   .templateField {
     width: 100%;
-    flex-direction: column;
-    align-items: flex-start;
+    flex-direction: row;
+    align-items: center;
     justify-content: center;
-    gap: 6px;
+    gap: 10px;
+    padding: 6px 0;
   }
 
-  .templateField a {
+  .templateNote {
     text-align: left;
-    word-break: break-word;
   }
 }
