+++ frontend/src/features/management/style/accountRegister.module.css
@@ -18,10 +18,10 @@
 
 .header {
   display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
+  flex-direction: column;
+  align-items: flex-start;
   width: 100%;
+  gap: 0.5rem;
 }
 
 .sectionName {
